添加将item放置在首页的功能

+++ app/src/main/java/an/devhp/manager/FragmentIds.java
@@ -9,17 +9,17 @@ package an.devhp.manager;
 
 public interface FragmentIds {
 
-    int HOME = 0x000;                       //    首页
+    long HOME = 0x000;                       //    首页
 
-    int SELECT_ANDROID_TECH = 0x101;        //    安卓选型
+    long SELECT_ANDROID_TECH = 0x101;        //    安卓选型
 
-    int SELECT_PICTURE = 0x102;             //    选择图片选型
+    long SELECT_PICTURE = 0x102;             //    选择图片选型
 
-    int SELECT_DB = 0x103;                  //    数据库选型
-    int SHOW_REALM_DB = 0x1031;             //    Realm show
+    long SELECT_DB = 0x103;                  //    数据库选型
+    long SHOW_REALM_DB = 0x1031;             //    Realm show
 
-    int SELECT_MEDIA = 0x104;               //    流媒体选型
-    int SHOW_EXO_PLAYER_MEDIA = 0x1041;     //    Exo Player show
+    long SELECT_MEDIA = 0x104;               //    流媒体选型
+    long SHOW_EXO_PLAYER_MEDIA = 0x1041;     //    Exo Player show
 
 
 }
