+++ app/src/main/java/an/devhp/ui/fragment/select/android/media/ShowExoPlayerFragment.java
@@ -74,7 +74,7 @@ public class ShowExoPlayerFragment extends SimpleListFragment {
     }
 
     @Override
-    public int getSimpleFragmentId() {
+    public long getSimpleFragmentId() {
         return FragmentIds.SHOW_EXO_PLAYER_MEDIA;
     }
