+++ app/src/main/java/an/devhp/ui/fragment/select/android/pic/SelectPicListFragment.java
@@ -28,7 +28,7 @@ public class SelectPicListFragment extends SimpleListFragment {
     }
 
     @Override
-    public int getSimpleFragmentId() {
+    public long getSimpleFragmentId() {
         return FragmentIds.SELECT_PICTURE;
     }
